fix TLog打日志打串的bug
